Get data from youtube api
I'm not using google's official youtube data api because it requires
OAUTH, and that is a bit "hacky" to set up.  Instead, I've taken the
necessary structs, and placed them here with my own http client.

If interested, here are some links:
  https://godoc.org/google.golang.org/api/youtube/v3 - docs for official
  client lib
  https://godoc.org/golang.org/x/oauth2 - docs for oauth
  https://developers.google.com/+/quickstart/go - sample golang client
  app for the google plus API (but uses OAUTH, so it's similiar)

In future, get rid of environment vars for something better